requirements.txt
requests==2.25.1
aiohttp==3.7.3
python-whois